remove home-screen widget, keep the Quick Settings tile
The widget added no capability the media notification didn't already do
better, while permanently occupying home-screen space in an app whose whole
character is immersive. The one thing that would have justified it —
changing the mix from the widget — was ruled out as too fragile during
design, so what shipped was a 4x1 bar with a name and a button.

The tile stays: it toggles playback from inside any other app without
leaving it, a capability the widget never had, and it has no visual design
to get wrong.

Removed the widget itself (AmbioWidget, AmbioWidgetReceiver, WidgetDisplay
and its test, widget_info.xml, the widget <receiver>), the Glance
dependencies it needed, and everything that existed only to feed it: the
four widget-only strings, AmbioApp's mix-change collector, and
AmbioAppViewModel.activeMix (the palette flow it sits beside predates the
widget and stays).

feature:widget is renamed to feature:tile, since it contains no widget
now and a module whose name states something false is exactly the kind of
leftover this cleanup is about. WidgetUpdater is renamed to PlaybackFlag,
since once its Glance-only members (currentDisplay, refresh, the Hilt
EntryPoint and the SoundRepository/StringProvider access they needed) are
gone, all it does is read and write one boolean flag for the tile to poll.
PlaybackStateReceiver no longer wraps that write in goAsync()/a
CoroutineScope — that machinery existed only because updating a Glance
widget used to suspend.

Verified via ./gradlew :feature:tile:compileDebugKotlin and
:app:compileDebugKotlin; full test/lint/assembleDebug gate left to the
next run.
