+++ app/build.gradle.kts
@@ -62,7 +62,7 @@ dependencies {
     implementation(project(":feature:home"))
     implementation(project(":feature:settings"))
     implementation(project(":feature:stats"))
-    implementation(project(":feature:widget"))
+    implementation(project(":feature:tile"))
     implementation(project(":media"))
     implementation(project(":ui"))
