+++ app/src/main/AndroidManifest.xml
@@ -44,11 +44,11 @@
         </service>
 
         <!-- Media3 routes ACTION_MEDIA_BUTTON here and starts AudioService if it is not
-             already running, which is what lets the widget and the tile control playback
-             hours after the app was last opened.
+             already running, which is what lets the tile control playback hours after the
+             app was last opened.
 
-             Declared here rather than in feature:widget, beside the AudioService it routes
-             to and whose declaration it depends on: it is app-wide plumbing, not widget
+             Declared here rather than in feature:tile, beside the AudioService it routes
+             to and whose declaration it depends on: it is app-wide plumbing, not tile
              plumbing. Headset and Bluetooth media buttons use this route too, so putting it
              in a feature module would mean unwiring that module silently removed
              hardware-button support from the whole app. media3-session's own AAR manifest
