+++ app/src/main/java/com/jbgsoft/ambio/AmbioApp.kt
@@ -1,9 +1,7 @@
 package com.jbgsoft.ambio
 
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
-import androidx.compose.ui.platform.LocalContext
 import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.compose.NavHost
@@ -12,7 +10,6 @@ import androidx.navigation.compose.rememberNavController
 import com.jbgsoft.ambio.feature.home.HomeScreen
 import com.jbgsoft.ambio.feature.settings.SettingsScreen
 import com.jbgsoft.ambio.feature.stats.StatsScreen
-import com.jbgsoft.ambio.feature.widget.WidgetUpdater
 import com.jbgsoft.ambio.ui.theme.AmbioTheme
 
 object Routes {
@@ -25,13 +22,6 @@ object Routes {
 fun AmbioApp(viewModel: AmbioAppViewModel = hiltViewModel()) {
     val palette by viewModel.palette.collectAsStateWithLifecycle()
     val navController = rememberNavController()
-
-    // The app is the only thing that knows the mix changed: the mix is written by this
-    // UI, and AudioService never sees an edit made while it is dead.
-    val context = LocalContext.current
-    LaunchedEffect(Unit) {
-        viewModel.activeMix.collect { WidgetUpdater.refresh(context) }
-    }
 
     // The theme wraps the whole graph, not just Home, so Settings and Stats
     // inherit the palette of the active mix.
