+++ app/src/main/java/com/jbgsoft/ambio/AmbioAppViewModel.kt
@@ -2,7 +2,6 @@ package com.jbgsoft.ambio
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.jbgsoft.ambio.core.domain.model.ActiveSound
 import com.jbgsoft.ambio.core.domain.model.AmbioPalette
 import com.jbgsoft.ambio.core.domain.model.SoundTheme
 import com.jbgsoft.ambio.core.domain.model.mixPalettes
@@ -26,17 +25,5 @@ class AmbioAppViewModel @Inject constructor(
             scope = viewModelScope,
             started = SharingStarted.WhileSubscribed(5_000),
             initialValue = SoundTheme.RAIN.toPalette()
-        )
-
-    /**
-     * The mix itself, from the same use case rather than derived from [palette]: two
-     * different mixes can share a palette but have different titles, and the widget's
-     * title has to follow the mix, not the colour.
-     */
-    val activeMix: StateFlow<List<ActiveSound>> = getActiveMix()
-        .stateIn(
-            scope = viewModelScope,
-            started = SharingStarted.WhileSubscribed(5_000),
-            initialValue = emptyList()
         )
 }
