+++ feature/tile/src/main/AndroidManifest.xml
@@ -19,17 +19,6 @@
         </service>
 
         <receiver
-            android:name=".AmbioWidgetReceiver"
-            android:exported="true">
-            <intent-filter>
-                <action android:name="android.appwidget.action.APPWIDGET_UPDATE" />
-            </intent-filter>
-            <meta-data
-                android:name="android.appwidget.provider"
-                android:resource="@xml/widget_info" />
-        </receiver>
-
-        <receiver
             android:name=".PlaybackStateReceiver"
             android:exported="false">
             <intent-filter>
