--- feature/widget/src/main/java/com/jbgsoft/ambio/feature/widget/AmbioWidgetReceiver.kt
@@ -1,8 +1,0 @@
-package com.jbgsoft.ambio.feature.widget
-
-import androidx.glance.appwidget.GlanceAppWidget
-import androidx.glance.appwidget.GlanceAppWidgetReceiver
-
-class AmbioWidgetReceiver : GlanceAppWidgetReceiver() {
-    override val glanceAppWidget: GlanceAppWidget = AmbioWidget
-}
