--- feature/widget/src/main/res/values/strings.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <string name="widget_sound_count">%1$d sounds</string>
-    <string name="widget_play">Play</string>
-    <string name="widget_pause">Pause</string>
-    <string name="widget_description">Play or pause your sound mix</string>
-</resources>
