--- feature/widget/src/main/res/xml/widget_info.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<appwidget-provider xmlns:android="http://schemas.android.com/apk/res/android"
-    android:description="@string/widget_description"
-    android:minWidth="250dp"
-    android:minHeight="50dp"
-    android:resizeMode="horizontal"
-    android:targetCellWidth="4"
-    android:targetCellHeight="1"
-    android:widgetCategory="home_screen"
-    android:initialLayout="@layout/glance_default_loading_layout" />
