+++ gradle/libs.versions.toml
@@ -6,7 +6,6 @@ ksp = "2.3.10"
 
 # Compose - Using BOM for version management
 compose-bom = "2026.06.01"
-glance = "1.1.1"
 
 # AndroidX
 hilt = "2.60.1"
@@ -74,10 +73,6 @@ room-compiler = { group = "androidx.room", name = "room-compiler", version.ref =
 
 # DataStore
 datastore-preferences = { group = "androidx.datastore", name = "datastore-preferences", version.ref = "datastore" }
-
-# Glance
-glance-appwidget = { group = "androidx.glance", name = "glance-appwidget", version.ref = "glance" }
-glance-material3 = { group = "androidx.glance", name = "glance-material3", version.ref = "glance" }
 
 # Coroutines
 kotlinx-coroutines-core = { group = "org.jetbrains.kotlinx", name = "kotlinx-coroutines-core", version.ref = "coroutines" }
