+++ settings.gradle.kts
@@ -31,6 +31,6 @@ include(":core:di")
 include(":feature:home")
 include(":feature:settings")
 include(":feature:stats")
-include(":feature:widget")
+include(":feature:tile")
 include(":media")
 include(":ui")
